% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_alignment.R
\name{choose_alignment}
\alias{choose_alignment}
\title{choose_alignment}
\usage{
choose_alignment(lib_path, lib_file, blast_file, align_path, alignment,
  viewinR = FALSE)
}
\arguments{
\item{lib_path}{string. path. location of strain library}

\item{lib_file}{string. file name. strain library file}

\item{blast_file}{string. file name. csv file of blast results}

\item{align_path}{string. path. location of alignment file}

\item{alignment}{string or vector of strings. file name. alignment file
downloaded from EMBL clustal omega.}

\item{viewinR}{logical. default \code{FALSE.} When TRUE, displays alignment
file in R console}
}
\value{
updates strain library with strains selected
}
\description{
function to choose sample based on alignment using EMBL clustal omega
}
