% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2fasta.R
\name{df2fasta}
\alias{df2fasta}
\title{df2fasta}
\usage{
df2fasta(df, fname = NULL, path, header_col, seq_col)
}
\arguments{
\item{df}{dataframe}

\item{fname}{string. name of fasta file}

\item{path}{string. location of output fasta files}

\item{header_col}{column name containing sequence headers}

\item{seq_col}{column name of sequences}
}
\description{
converts dataframe into fasta file, where fasta headers are in one dataframe
column and sequences are in another dataframe column
}
