% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_strainlib.R
\name{gen_strainlib}
\alias{gen_strainlib}
\title{gen_strainlib}
\usage{
gen_strainlib(blast_file, lib_path, lib_file = NULL,
  lib_name = "mystrainlibrary")
}
\arguments{
\item{blast_file}{string. file name. csv file containing blast results. Include path if necessary}

\item{lib_path}{path. string. location of strain library file}

\item{lib_file}{string. file name. defualt NULL. Name of strain library file.
Creates a library file if none supplied.}

\item{lib_name}{string. defaults to generic library name,name of strain library (only used if new one is
being created)}
}
\value{
creates/updates strain library
    creates fasta file of strain library
    creates fasta files of species with multiple isolates found
}
\description{
function to generate/update strain library and fasta file
reads in blast result file from parse blast
if no strain library supplied, creates a new one
Species that are only found once and not already in strain library are
added to library
Generates strain library as csv
Generates strain library as fasta
also generates fasta file for each species, amalgamating results from
current blast result and strain library
}
