% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_alignment.R
\name{choose_alignment}
\alias{choose_alignment}
\title{choose_alignment}
\usage{
choose_alignment(lib_path, lib_file, blast_file, fasta_path, fastatoalign)
}
\arguments{
\item{lib_path}{string. path. location of strain library}

\item{lib_file}{string. file name. strain library file}

\item{blast_file}{string. file name. csv file of blast results}

\item{fasta_path}{string. path. location of fasta file(s) that need to be aligned}

\item{fastatoalign}{string or vector of strings. fasta files to be aligned.}
}
\value{
updates strain library with strains selected
}
\description{
function to choose sample based on alignment using EMBL clustal omega
}
